requirements.txt
pycryptodome
msgpack